Refactor: coupon status tag component table cell page 에 병합
