--- src/components/domain/coupon/coupon-status-tag/type.ts
@@ -1,8 +1,0 @@
-export type StyledCouponTagProps = {
-  backgroundColor: string;
-  borderColor: string;
-  color: string;
-};
-export type CouponTagProps = {
-  status: string;
-};
